requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
django-crispy-forms
django-simple-captcha